setup sdk resources as classes
