--- src/resources/Bounty/constants.js
@@ -1,25 +1,0 @@
-export const DIFFICULTY_VALUES = {
-  Beginner: 0,
-  Intermediate: 1,
-  Expert: 2
-};
-
-export const DIFFICULTY_MAPPINGS = {
-  0: 'Beginner',
-  1: 'Intermediate',
-  2: 'Expert'
-};
-
-export const STAGE_VALUES = {
-  0: 'Draft',
-  1: 'Active',
-  2: 'Dead',
-  3: 'Completed',
-  4: 'Expired'
-};
-
-export const DRAFT = 0;
-export const ACTIVE = 1;
-export const DEAD = 2;
-export const COMPLETED = 3;
-export const EXPIRED = 4;
