--- src/resources/Fulfillment/selectors.js
@@ -1,8 +1,0 @@
-import { createSelector } from 'reselect';
-
-export const rootFulfillmentSelector = state => state.fulfillment;
-
-export const fulfillmentSelector = createSelector(
-  rootFulfillmentSelector,
-  rootFulfillment => rootFulfillment
-);
