--- src/resources/Fulfillments/constants.js
@@ -1,1 +1,0 @@
-export const LIMIT = 10;
