--- src/utils/Web3Client.js
@@ -1,7 +1,0 @@
-import Web3 from 'web3';
-
-const web3 = new Web3(
-  new Web3.providers.HttpProvider('https://mainnet.infura.io')
-);
-
-export default web3;
